+++ ChangeLog.md
@@ -1,6 +1,8 @@
 #2.2.9
   * Use SSL-enabled server for shortlinks
   * Add falloff for weapons
+  * Use falloff when calculating weapon effectiveness in damage dealt
+  * Add engagement range slider to allow user to see change in weapon effectiveness with range
 
 #2.2.8
   * Fix issue where filling all internals with cargo racks would include restricted slots